requirements.txt
aws-cdk-lib==2.20.0
constructs>=10.0.0,<11.0.0
black==22.3.0
boto3==1.21.42
